changed steps to long long

+++ josh.h
@@ -8,9 +8,9 @@
 
 void printArrays(int asc[], int des[], int ran[], int size);
 
-void bubbleSort(int arr[], int n, int& c);
+void bubbleSort(int arr[], int n, long long int& c);
 
-void merge(int arr[], int left, int mid, int right, int& c);
-void mergeSort(int arr[], int begin, int end, int& pass);
+void merge(int arr[], int left, int mid, int right, long long int& c);
+void mergeSort(int arr[], int begin, int end, long long int& pass);
